orchestration/requirements.txt
# =============================================================================
# Airflow Extra Dependencies — NYC TLC Pipeline
#
# Installed on top of the base apache/airflow:2.9.3 image.
# All direct and conflict-prone transitive dependencies are pinned for
# reproducible builds. Upgrade intentionally — do not use open ranges.
#
# Known constraints:
#   - protobuf must be 4.x: dbt-core 1.8.x requires MessageToJson(
#     always_print_fields_with_no_presence) added in protobuf 4.x.
#     Airflow 2.9.3 allows protobuf 4.x (!=4.0.*,!=4.21.*,<5.0).
#   - astronomer-cosmos [dbt-snowflake] extra omitted: dbt-snowflake is
#     installed explicitly in the step above, avoiding ResolutionTooDeep.
# =============================================================================

# --- Snowflake provider ---
apache-airflow-providers-snowflake==5.6.0

# --- dbt ---
dbt-core==1.8.7
dbt-snowflake==1.8.4

# --- Astronomer Cosmos (Airflow <> dbt bridge) ---
astronomer-cosmos==1.8.0

# --- Protobuf: force 4.x for dbt-core 1.8.x compatibility ---
protobuf==4.25.3

# --- Azure Blob Storage (download_to_azure DAG task) ---
azure-storage-blob==12.22.0
requests==2.32.3

# --- MLflow (mlflow_cleanup DAG + future ML DAGs) ---
mlflow==2.19.0

# --- ML model dependencies ---
lightgbm==4.5.0
pandas==2.1.4
numpy==1.26.4
scikit-learn==1.5.2
matplotlib==3.9.2
scipy==1.13.1
statsmodels==0.14.4
snowflake-connector-python[pandas]==3.12.2
pyarrow==18.1.0
